Refactor property rules
