+++ plugins/@grouparoo/demo/src/sample_data.ts
@@ -1,7 +1,13 @@
 import Database from "./util/postgres";
 import { log, init } from "./util/shared";
 import { runAction } from "./util/runAction";
-import { App, Source, ProfilePropertyRule, Schedule } from "@grouparoo/core";
+import {
+  makePropertyRuleIdentifying,
+  createPropertyRule,
+  ensureBootstrapPropertyRule,
+  RuleDefinition,
+} from "./util/propertyRules";
+import { App, Source, Schedule } from "@grouparoo/core";
 
 export const SCHEMA_NAME = "demo";
 export const USER_ID_PROPERTY_NAME = "userId";
@@ -18,7 +24,7 @@ const USERS = {
   deactivated: "BOOLEAN",
 };
 
-const USER_ID_PROPERTY_RULE = {
+const USER_ID_PROPERTY_RULE: RuleDefinition = {
   key: USER_ID_PROPERTY_NAME,
   type: "integer",
   unique: true,
@@ -26,14 +32,7 @@ const USER_ID_PROPERTY_RULE = {
   options: { column: "id", "aggregation method": "exact" },
 };
 
-const RULE_DEFAULT = {
-  isArray: false,
-  unique: false,
-  identifying: false,
-  filters: [],
-};
-
-const USER_RULES = [
+const USER_RULES: Array<RuleDefinition> = [
   {
     key: "firstName",
     type: "string",
@@ -72,7 +71,7 @@ const PURCHASES = {
 };
 
 const PURCHASE_FILTERS = [{ key: "state", op: "equals", match: "successful" }];
-const PURCHASE_RULES = [
+const PURCHASE_RULES: Array<RuleDefinition> = [
   {
     key: "LTV",
     type: "float",
@@ -114,7 +113,7 @@ export async function users() {
   await createCsvTable("users", "id", USERS, true, true);
   await createSource("users", "id");
   await createPropertyRules("users", USER_RULES);
-  await makePropertyRuleIdentifying("users", "email");
+  await setEmailAsIdentifying("users");
   await createAndRunSchedule("users", "updated_at");
 }
 
@@ -180,93 +179,6 @@ async function findTableSource(app, tableName) {
   return null;
 }
 
-async function makePropertyRuleIdentifying(
-  tableName: string,
-  propertyKey: string
-) {
-  // remove current one
-  const current = await ProfilePropertyRule.scope(null).findAll({
-    where: { identifying: true },
-  });
-  for (const propertyRule of current) {
-    const params = { identifying: false, guid: propertyRule.guid };
-    await runAction("profilePropertyRule:edit", params);
-  }
-
-  // set on new one
-  const source = await findSource(tableName);
-  const found = await findPropertyRule(source, propertyKey);
-  if (!found) {
-    throw new Error(`Identifying rule not found (${propertyKey})`);
-  }
-
-  const params = { identifying: false, guid: found.guid };
-  await runAction("profilePropertyRule:edit", params);
-}
-
-async function findPropertyRule(source: Source, propertyKey: string) {
-  const where = {
-    sourceGuid: source.guid,
-    key: propertyKey,
-  };
-  return ProfilePropertyRule.scope(null).findOne({ where });
-}
-
-async function createPropertyRules(tableName: string, rules: any) {
-  const source = await findSource(tableName);
-  for (const rule of rules) {
-    await createPropertyRule(source, rule);
-  }
-}
-
-async function createPropertyRule(source: Source, rule: any) {
-  const found = await findPropertyRule(source, rule.key);
-
-  const params = Object.assign({}, RULE_DEFAULT, rule, {
-    state: "ready",
-    sourceGuid: source.guid,
-    guid: found?.guid,
-  });
-  if (found) {
-    await runAction("profilePropertyRule:edit", params);
-  } else {
-    await runAction("profilePropertyRule:create", params);
-  }
-  const made = await findPropertyRule(source, rule.key);
-  if (!made) {
-    throw new Error(`rule not made: ${rule.key}`);
-  }
-  return made;
-}
-
-async function ensureUserIdPropertyRule(source: Source) {
-  const rule = USER_ID_PROPERTY_RULE;
-  const found = await findPropertyRule(source, rule.key);
-  if (found) {
-    const params = Object.assign({}, RULE_DEFAULT, rule, {
-      state: "ready",
-      sourceGuid: source.guid,
-      guid: found?.guid,
-    });
-    await runAction("profilePropertyRule:edit", params);
-  } else {
-    // need to bootstrap it
-    const params = {
-      guid: source.guid,
-      key: rule.key,
-      type: rule.type,
-      mappedColumn: rule.options.column,
-    };
-    await runAction("source:bootstrapUniqueProfilePropertyRule", params);
-  }
-
-  const made = await findPropertyRule(source, rule.key);
-  if (!made) {
-    throw new Error("unique userId not made app not created!");
-  }
-  return made;
-}
-
 async function findSource(tableName: string) {
   const app = await getApp();
   const found = await findTableSource(app, tableName);
@@ -304,7 +216,7 @@ async function getSource(app: App, tableName: string, userId: string) {
     throw new Error(`Source not created (${tableName})!`);
   }
   if (tableName === "users") {
-    await ensureUserIdPropertyRule(made);
+    await ensureBootstrapPropertyRule(made, USER_ID_PROPERTY_RULE);
   }
   const update = {
     guid: made.guid,
@@ -316,6 +228,21 @@ async function getSource(app: App, tableName: string, userId: string) {
 
   await made.reload();
   return made;
+}
+
+export async function createPropertyRules(
+  tableName: string,
+  rules: Array<RuleDefinition>
+) {
+  const source = await findSource(tableName);
+  for (const rule of rules) {
+    await createPropertyRule(source, rule);
+  }
+}
+
+async function setEmailAsIdentifying(tableName: string) {
+  const source = await findSource(tableName);
+  await makePropertyRuleIdentifying(source, "email");
 }
 
 async function getApp() {
